refactor(js): relocate `budget-collapse` module to `views/pages/budget-plan`
